+++ src/App.css
@@ -303,3 +303,10 @@ button[type=submit]{
   border-radius: 4px;
   cursor: pointer;
 }
+
+/*about us*/
+.about{
+  text-align: center;
+  color: blue;
+  width: auto;
+}
