+++ src/components/home/about.js
@@ -1,0 +1,25 @@
+import React, { Component } from 'react'
+import './App.css'
+
+ class About extends Component {
+    render() {
+        return (
+            <div className ='about'>
+                <h1>Welcome</h1><hr/>
+                <h2>~~~ Vision Statement ~~~</h2>
+                <p>Bacon ipsum dolor sit amet salami ham hock ham</p>
+                <h2>~~~ Mission Statement ~~~</h2>
+                  <p>Bacon ipsum dolor sit amet salami ham hock ham</p><hr/>
+                <h2>~~~ Our Story ~~~</h2>
+                <article>
+                   
+                    <article>
+                   
+                    </article>
+                </article>
+                
+            </div>
+        )
+    }
+}
+export default About;
